change to passing a Domain list around rather than a map of domain/emails.
