+++ lib/shelf_letsencrypt.dart
@@ -2,4 +2,5 @@
 library shelf_letsencrypt;
 
 export 'src/certs_handler.dart';
+export 'src/domain.dart';
 export 'src/letsencrypt.dart';
