+++ lib/src/certs_handler.dart
@@ -2,11 +2,11 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:basic_utils/basic_utils.dart';
+import 'package:basic_utils/basic_utils.dart' hide Domain;
 import 'package:collection/collection.dart';
-import 'package:path/path.dart' as pack_path;
 
 import '../shelf_letsencrypt.dart';
+import 'pem_key_pair.dart';
 
 /// Base class for a certificate handler.
 ///
@@ -44,7 +44,7 @@ abstract class CertificatesHandler {
   /// for [domains] it will return `null`.
   ///
   /// See [LetsEncrypt.startSecureServer].
-  FutureOr<SecurityContext?> buildSecurityContext(List<String> domains,
+  FutureOr<SecurityContext?> buildSecurityContext(List<Domain> domains,
       {bool loadAllHandledDomains = true});
 
   /// Returns a [List] of all the handled domains.
@@ -63,10 +63,10 @@ abstract class CertificatesHandler {
           .toList();
 
   /// Returns a [List] of NOT handled [domains].
-  List<String> listNotHandledDomains(List<String> domains,
+  List<Domain> listNotHandledDomains(List<Domain> domains,
           {bool checkSecurityContext = true}) =>
       domains
-          .where((d) => !isHandledDomainCertificate(d,
+          .where((d) => !isHandledDomainCertificate(d.name,
               checkSecurityContext: checkSecurityContext))
           .toList();
 
@@ -252,29 +252,6 @@ abstract class CertificatesHandler {
   }
 }
 
-/// A [SecurityContext] builder.
-class SecurityContextBuilder {
-  static bool defineMerged = false;
-
-  final Set<DomainCertificate> domainsCertificates = <DomainCertificate>{};
-
-  SecurityContext build() {
-    final securityContext = SecurityContext();
-
-    if (defineMerged) {
-      domainsCertificates
-          .reduce((value, element) => value.merge(element))
-          .define(securityContext);
-    } else {
-      for (final d in domainsCertificates) {
-        d.define(securityContext);
-      }
-    }
-
-    return securityContext;
-  }
-}
-
 /// Holds [domains] certificates to load into a [SecurityContext].
 abstract class DomainCertificate {
   DomainCertificate(Iterable<String> domains) {
@@ -335,334 +312,4 @@ class DomainCertificatePEM extends DomainCertificate {
       ..useCertificateChainBytes(utf8.encode(fullChainPEM))
       ..usePrivateKeyBytes(utf8.encode(privateKeyPEM));
   }
-}
-
-/// A [DomainCertificate] implementation using file paths.
-class DomainCertificateFilePath extends DomainCertificate {
-  DomainCertificateFilePath(List<String> super.domains, this.fullChainFilePath,
-      this.privateKeyFilePath);
-  final String fullChainFilePath;
-
-  final String privateKeyFilePath;
-
-  @override
-  String get fullChainPEM => File(fullChainFilePath).readAsStringSync();
-
-  @override
-  String get privateKeyPEM => File(privateKeyFilePath).readAsStringSync();
-
-  @override
-  void define(SecurityContext securityContext) {
-    securityContext
-      ..useCertificateChain(fullChainFilePath)
-      ..usePrivateKey(privateKeyFilePath);
-  }
-}
-
-/// A [CertificatesHandler] implementation using [dart:io].
-class CertificatesHandlerIO extends CertificatesHandler {
-  CertificatesHandlerIO(this.directory,
-      {super.accountDirectory,
-      super.privateKeyPEMFileName,
-      super.publicKeyPEMFileName,
-      super.fullChainPEMFileName}) {
-    directory.createSync(recursive: true);
-  }
-
-  /// The [Directory] to storage certificates.
-  final Directory directory;
-
-  @override
-  List<String> listAllHandledDomains({bool checkSecurityContext = true}) {
-    final domainsDirs = directory
-        .listSync()
-        .where(_isDomainDirectory)
-        .map((e) => _pathFileName(e.path))
-        .where((e) => isHandledDomainCertificate(e,
-            checkSecurityContext: checkSecurityContext))
-        .toList();
-
-    return domainsDirs;
-  }
-
-  bool _isDomainDirectory(FileSystemEntity e) {
-    if (e.statSync().type != FileSystemEntityType.directory) {
-      return false;
-    }
-
-    final dirName = _pathFileName(e.path);
-
-    if (dirName == accountDirectory || dirName.startsWith('.')) {
-      return false;
-    }
-
-    final dir = Directory(e.path);
-
-    final files = dir.listSync(followLinks: false);
-    final pemFiles = files.where((f) => f.path.endsWith('pem')).toList();
-
-    if (pemFiles.isEmpty) {
-      return false;
-    }
-
-    final fullChainFile =
-        pemFiles.firstWhereOrNull((f) => f.path.endsWith(fullChainPEMFileName));
-
-    if (fullChainFile == null) {
-      return false;
-    }
-
-    final privateKeyFile = pemFiles
-        .firstWhereOrNull((f) => f.path.endsWith(privateKeyPEMFileName));
-
-    if (privateKeyFile == null) {
-      return false;
-    }
-
-    return true;
-  }
-
-  String _pathFileName(String path) => pack_path.split(path).last;
-
-  @override
-  Future<SecurityContext?> buildSecurityContext(List<String> domains,
-      {bool loadAllHandledDomains = true}) async {
-    final securityContextBuilder = SecurityContextBuilder();
-
-    for (final domain in domains) {
-      final domainOk =
-          await _useDomainCertificate(securityContextBuilder, domain);
-      if (!domainOk) {
-        return null;
-      }
-    }
-
-    if (loadAllHandledDomains) {
-      final handledDomains = listAllHandledDomains();
-
-      for (final d in handledDomains) {
-        if (!domains.contains(d)) {
-          await _useDomainCertificate(securityContextBuilder, d);
-        }
-      }
-    }
-
-    final securityContext = securityContextBuilder.build();
-    return securityContext;
-  }
-
-  Future<bool> _useDomainCertificate(
-      SecurityContextBuilder securityContextBuilder, String domain) async {
-    final fullChainFile = fileDomainFullChainPEM(domain);
-    final privateKeyFile = fileDomainPrivateKeyPEM(domain);
-
-    if (!_fileExistsWithContent(fullChainFile) ||
-        !_fileExistsWithContent(privateKeyFile)) {
-      return false;
-    }
-
-    final fullChainPath = fullChainFile.path;
-    final privateKeyPath = privateKeyFile.path;
-
-    securityContextBuilder.domainsCertificates.add(
-        DomainCertificateFilePath([domain], fullChainPath, privateKeyPath));
-
-    return true;
-  }
-
-  @override
-  bool isHandledDomainCertificate(String domain,
-      {bool checkSecurityContext = true}) {
-    final fullChainFile = fileDomainFullChainPEM(domain);
-    final privateKeyFile = fileDomainPrivateKeyPEM(domain);
-
-    if (!_fileExistsWithContent(fullChainFile) ||
-        !_fileExistsWithContent(privateKeyFile)) {
-      return false;
-    }
-
-    final certificateExpired =
-        isCertificateExpired(fullChainFile.readAsStringSync());
-    if (certificateExpired) {
-      return false;
-    }
-
-    if (!checkSecurityContext) {
-      return true;
-    }
-
-    try {
-      final fullChainPath = fullChainFile.path;
-      final privateKeyPath = privateKeyFile.path;
-
-      SecurityContext()
-        ..useCertificateChain(fullChainPath)
-        ..usePrivateKey(privateKeyPath);
-
-      return true;
-    // ignore: avoid_catches_without_on_clauses
-    } catch (_) {
-      return false;
-    }
-  }
-
-  File fileAccountPrivateKeyPEM() => File(
-      pack_path.join(directory.path, accountDirectory, privateKeyPEMFileName));
-
-  File fileAccountPublicKeyPEM() => File(
-      pack_path.join(directory.path, accountDirectory, publicKeyPEMFileName));
-
-  File fileDomainPrivateKeyPEM(String cn) =>
-      File(pack_path.join(directory.path, cn, privateKeyPEMFileName));
-
-  File fileDomainPublicKeyPEM(String cn) =>
-      File(pack_path.join(directory.path, cn, publicKeyPEMFileName));
-
-  File fileDomainFullChainPEM(String cn) =>
-      File(pack_path.join(directory.path, cn, fullChainPEMFileName));
-
-  @override
-  FutureOr<String?> loadAccountPrivateKeyPEM() {
-    final file = fileAccountPrivateKeyPEM();
-    return _readFileWithContent(file);
-  }
-
-  @override
-  FutureOr<String?> loadAccountPublicKeyPEM() {
-    final file = fileAccountPublicKeyPEM();
-    return _readFileWithContent(file);
-  }
-
-  @override
-  FutureOr<String?> loadDomainPrivateKeyPEM(String cn) {
-    final file = fileDomainPrivateKeyPEM(cn);
-    return _readFileWithContent(file);
-  }
-
-  @override
-  FutureOr<String?> loadDomainPublicKeyPEM(String cn) {
-    final file = fileDomainPublicKeyPEM(cn);
-    return _readFileWithContent(file);
-  }
-
-  String? _readFileWithContent(File file) {
-    if (_fileExistsWithContent(file)) {
-      return file.readAsStringSync();
-    }
-    return null;
-  }
-
-  bool _fileExistsWithContent(File file) =>
-      file.existsSync() && file.lengthSync() > 1;
-
-  @override
-  Future<PEMKeyPair?> generateAccountPEMKeyPair({bool force = false}) async {
-    final filePrivateKey = fileAccountPrivateKeyPEM();
-    final filePublicKey = fileAccountPublicKeyPEM();
-
-    return _generateStoredKeyPair(filePrivateKey, filePublicKey, force);
-  }
-
-  @override
-  Future<PEMKeyPair?> generateDomainPEMKeyPair(String cn,
-      {bool force = false}) async {
-    final filePrivateKey = fileDomainPrivateKeyPEM(cn);
-    final filePublicKey = fileDomainPublicKeyPEM(cn);
-
-    return _generateStoredKeyPair(filePrivateKey, filePublicKey, force);
-  }
-
-  Future<PEMKeyPair?> _generateStoredKeyPair(
-      File filePrivateKey, File filePublicKey, bool force) async {
-    if (!force &&
-        (_fileExistsWithContent(filePrivateKey) ||
-            _fileExistsWithContent(filePublicKey))) {
-      return getAccountPEMKeyPair();
-    }
-
-    final keyPair = generatePEMKeyPair();
-
-    filePrivateKey.parent.createSync(recursive: true);
-    filePublicKey.parent.createSync(recursive: true);
-
-    filePrivateKey.writeAsStringSync(keyPair.privateKeyPEM);
-    filePublicKey.writeAsStringSync(keyPair.publicKeyPEM);
-
-    return keyPair;
-  }
-
-  @override
-  PEMKeyPair generatePEMKeyPair() {
-    final rsaKeyPair = CryptoUtils.generateRSAKeyPair();
-
-    final rsaPrivateKey = rsaKeyPair.privateKey as RSAPrivateKey;
-    final rsaPublicKey = rsaKeyPair.publicKey as RSAPublicKey;
-
-    final privateKeyPEM = CryptoUtils.encodeRSAPrivateKeyToPem(rsaPrivateKey);
-    final publicKeyPEM = CryptoUtils.encodeRSAPublicKeyToPem(rsaPublicKey);
-
-    final keyPair =
-        PEMKeyPair(privateKeyPEM, publicKeyPEM, rsaPrivateKey, rsaPublicKey);
-    return keyPair;
-  }
-
-  @override
-  Future<String?> generateCSR(String cn, String email,
-      {String? organizationName,
-      String? organizationUnit,
-      String? locality,
-      String? state,
-      String? country}) async {
-    final domainKeyPair = await getDomainPEMKeyPair(cn);
-    if (domainKeyPair == null) {
-      return null;
-    }
-
-    final attributes = {
-      'CN': cn,
-      if (organizationName != null) 'O': organizationName,
-      if (organizationUnit != null) 'OU': organizationUnit,
-      if (locality != null) 'L': locality,
-      if (state != null) 'ST': state,
-      if (country != null) 'C': country,
-    };
-
-    final csr = X509Utils.generateRsaCsrPem(
-        attributes, domainKeyPair.privateKey, domainKeyPair.publicKey);
-
-    return csr;
-  }
-
-  @override
-  Future<bool> saveSignedCertificateChain(
-      String cn, List<String> signedCertificatesChain) async {
-    final fullChainPEM = CertificatesHandler.joinPEMs(signedCertificatesChain);
-
-    final filePrivateKey = fileDomainFullChainPEM(cn);
-
-    filePrivateKey.parent.createSync(recursive: true);
-    filePrivateKey.writeAsStringSync(fullChainPEM);
-
-    return true;
-  }
-
-  @override
-  String toString() => 'CertificatesHandlerIO@$directory';
-}
-
-class PEMKeyPair {
-  PEMKeyPair(this.privateKeyPEM, this.publicKeyPEM,
-      [this._privateKey, this._publicKey]);
-  final String privateKeyPEM;
-  final String publicKeyPEM;
-
-  RSAPrivateKey? _privateKey;
-
-  RSAPrivateKey get privateKey =>
-      _privateKey ??= CryptoUtils.rsaPrivateKeyFromPem(privateKeyPEM);
-
-  RSAPublicKey? _publicKey;
-
-  RSAPublicKey get publicKey =>
-      _publicKey ??= CryptoUtils.rsaPublicKeyFromPem(publicKeyPEM);
 }
